added other green path test method for checkinthe blacklits
delete service retuning 500 when there is no ip range exists in the database

+++ src/main/java/com/finreach/codechallenge/filteringsystem/filteringsystem/service/FilterConfigurationService.java
@@ -68,6 +68,7 @@ public class FilterConfigurationService {
 
     public List<FilterConfigurationDTO> deleteFilterConfigurationList(FilterConfigurationDTO filterConfigurationDTO) {
         List<FilterConfiguration> filterConfigurationListToBeDeleted = filterConfigurationRepository.findByFromIPAndToIP(filterConfigurationDTO.getFromIP(), filterConfigurationDTO.getToIP());
+        if (filterConfigurationListToBeDeleted.size()==0) throw new RuntimeException("This ip range does not exist in the black list");
         filterConfigurationListToBeDeleted.stream().forEach(filterConfiguration -> {
             filterConfigurationRepository.delete(filterConfiguration);
         });
